backend/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
pydantic
pydantic-settings
psycopg2-binary
python-dotenv
python-jose
python-jose[cryptography]
opencv-python-headless
qrcode 
numpy 
pillow
python-multipart
opencv-python
passlib[bcrypt]
pydantic[email]
pandas
openpyxl
requests
boto3
razorpay>=1.4.0,<1.5.0